// lib/types.ts
export type Role =
  | "customer"
  | "vendor_owner"
  | "outlet_manager"
  | "admin"
  | "approver"
  | "super_admin";

export interface User {
  id: string;
  name: string;
  email: string;
  role: Role;
  avatarInitial: string;
  city?: string;
  verificationTier: "guest" | "registered" | "phone_verified" | "profile_complete" | "kyc_verified";
  vendorId?: string; // set for vendor_owner
  outletId?: string; // set for outlet_manager
}